Revert "hima: audioamp: Stupid 64bit..."
This reverts commit 493ed174f743bdaf92f9fba78e04126aca786aec.

Conflicts:
	libaudioamp/tfa9887.cpp

+++ amplifier/tfa9895.c
@@ -19,7 +19,6 @@
 
 #include <errno.h>
 #include <fcntl.h>
-#include <stdint.h>
 #include <stdio.h>
 #include <string.h>
 #include <unistd.h>
@@ -275,7 +274,7 @@ static int tfa9895_read_reg(struct tfa9895_amp_t *amp, uint8_t reg,
     }
 
     reg_val[0] = 2;
-    reg_val[1] = (uint64_t) &buf;
+    reg_val[1] = (unsigned int) &buf;
     /* unsure why the first byte is skipped */
     buf[0] = 0;
     buf[1] = reg;
@@ -285,7 +284,7 @@ static int tfa9895_read_reg(struct tfa9895_amp_t *amp, uint8_t reg,
     }
 
     reg_val[0] = 2;
-    reg_val[1] = (uint64_t) &buf;
+    reg_val[1] = (unsigned int) &buf;
     if ((ret = ioctl(amp->fd, TPA9895_READ_CONFIG, &reg_val)) != 0) {
         ALOGE("ioctl %d failed, ret = %d", TPA9895_READ_CONFIG, ret);
         goto read_reg_err;
@@ -310,7 +309,7 @@ static int tfa9895_write_reg(struct tfa9895_amp_t *amp, uint8_t reg,
     }
 
     reg_val[0] = 4;
-    reg_val[1] = (uint64_t) &buf;
+    reg_val[1] = (unsigned int) &buf;
     /* unsure why the first byte is skipped */
     buf[0] = 0;
     buf[1] = reg;
@@ -339,7 +338,7 @@ static int tfa9895_read(struct tfa9895_amp_t *amp, int addr, uint8_t *buf,
     }
 
     reg_val[0] = 2;
-    reg_val[1] = (uint64_t) &reg_buf;
+    reg_val[1] = (unsigned int) &reg_buf;
     /* unsure why the first byte is skipped */
     reg_buf[0] = 0;
     reg_buf[1] = (0xFF & addr);
@@ -349,7 +348,7 @@ static int tfa9895_read(struct tfa9895_amp_t *amp, int addr, uint8_t *buf,
     }
 
     reg_val[0] = len;
-    reg_val[1] = (uint64_t) &kernel_buf;
+    reg_val[1] = (unsigned int) &kernel_buf;
     if ((ret = ioctl(amp->fd, TPA9895_READ_CONFIG, &reg_val)) != 0) {
         ALOGE("ioctl %d failed, ret = %d", TPA9895_READ_CONFIG, ret);
         goto read_err;
@@ -373,7 +372,7 @@ static int tfa9895_write(struct tfa9895_amp_t *amp, int addr,
     }
 
     reg_val[0] = len + 2;
-    reg_val[1] = (uint64_t) &ioctl_buf;
+    reg_val[1] = (unsigned int) &ioctl_buf;
     /* unsure why the first byte is skipped */
     ioctl_buf[0] = 0;
     ioctl_buf[1] = (0xFF & addr);
